refactor: api.ts에서 http 분리
