+++ src/api.ts
@@ -1,200 +1,17 @@
-import { SERVER_URL } from "@/components/const";
-import { Folder, Product, User } from "@/const";
+import {
+  ApiGetFunction,
+  ApiModifyFunction,
+  FailCallback,
+  SucceedResponse,
+  SuccessCallback,
+} from "@/api/const";
+import http from "@/api/http";
+import { deleteProduct, putProduct } from "@/api/products";
 import {
   transformProductForFolderUpdate,
   transformUserForUpdate,
-} from "@/util";
-
-const API_VERSION = "/v1";
-
-interface SucceedResponse {
-  message: string;
-}
-
-interface ErrorResponse {
-  error: string;
-}
-
-type SuccessCallback<T> = (data: T) => void;
-type FailCallback = (error: Error) => void;
-
-type ApiGetFunction<T> = (
-  onSuccess: SuccessCallback<T>,
-  onFail: FailCallback,
-  targetId?: string
-) => Promise<void>;
-
-type ApiModifyFunction<T> = (
-  body: BodyInit | null | undefined,
-  onSuccess: SuccessCallback<T>,
-  onFail: FailCallback,
-  targetId?: string
-) => Promise<void>;
-
-const isErrorResponse = (data: unknown): data is ErrorResponse => {
-  return typeof data === "object" && data !== null && "error" in data;
-};
-
-const handleResponse = <T>(res: Response): Promise<T> =>
-  res.json().then((data: T | ErrorResponse) => {
-    if (!res.ok && isErrorResponse(data)) {
-      throw new Error(data.error);
-    }
-    return data as T;
-  });
-
-const http = {
-  get: <T>(
-    path: string,
-    options = {},
-    onSuccess: SuccessCallback<T>,
-    onFail: FailCallback
-  ) =>
-    fetch(`${SERVER_URL}${API_VERSION}${path}`, {
-      method: "GET",
-      ...options,
-    })
-      .then((res) => handleResponse<T>(res))
-      .then(onSuccess)
-      .catch(onFail),
-  post: <T>(
-    path: string,
-    options = {},
-    body: BodyInit | null | undefined,
-    onSuccess: SuccessCallback<T>,
-    onFail: FailCallback
-  ) =>
-    fetch(`${SERVER_URL}${API_VERSION}${path}`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      ...options,
-      body,
-    })
-      .then((res) => handleResponse<T>(res))
-      .then(onSuccess)
-      .catch(onFail),
-  put: <T>(
-    path: string,
-    options = {},
-    body: BodyInit | null | undefined,
-    onSuccess: SuccessCallback<T>,
-    onFail: FailCallback
-  ) =>
-    fetch(`${SERVER_URL}${API_VERSION}${path}`, {
-      method: "PUT",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      ...options,
-      body,
-    })
-      .then((res) => handleResponse<T>(res))
-      .then(onSuccess)
-      .catch(onFail),
-  patch: <T>(
-    path: string,
-    options = {},
-    body: BodyInit | null | undefined,
-    onSuccess: SuccessCallback<T>,
-    onFail: FailCallback
-  ) =>
-    fetch(`${SERVER_URL}${API_VERSION}${path}`, {
-      method: "PATCH",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      ...options,
-      body,
-    })
-      .then((res) => handleResponse<T>(res))
-      .then(onSuccess)
-      .catch(onFail),
-  delete: <T>(
-    path: string,
-    options = {},
-    body: BodyInit | null | undefined,
-    onSuccess: SuccessCallback<T>,
-    onFail: FailCallback
-  ) =>
-    fetch(`${SERVER_URL}${API_VERSION}${path}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      ...options,
-      body,
-    })
-      .then((res) => handleResponse<T>(res))
-      .then(onSuccess)
-      .catch(onFail),
-};
-
-export const getProductList: ApiGetFunction<Product[]> = (
-  onSuccess,
-  onFail
-) => {
-  return http.get(`/products`, undefined, onSuccess, onFail);
-};
-
-export const postProduct: ApiModifyFunction<SucceedResponse> = (
-  body,
-  onSuccess,
-  onFail
-) => {
-  return http.post(
-    `/products`,
-    { credentials: "include", headers: {} },
-    body,
-    onSuccess,
-    onFail
-  );
-};
-
-export const postProducts: ApiModifyFunction<SucceedResponse> = (
-  body,
-  onSuccess,
-  onFail
-) => {
-  return http.post(
-    `/products/batch`,
-    { credentials: "include" },
-    body,
-    onSuccess,
-    onFail
-  );
-};
-
-export const putProduct: ApiModifyFunction<Product> = (
-  body,
-  onSuccess,
-  onFail,
-  targetId
-) => {
-  return http.put(
-    `/products/${targetId}`,
-    { credentials: "include", headers: {} },
-    body,
-    onSuccess,
-    onFail
-  );
-};
-
-const deleteProduct: ApiModifyFunction<SucceedResponse> = (
-  body,
-  onSuccess,
-  onFail,
-  targetId
-) => {
-  return http.delete(
-    `/products/${targetId}`,
-    { credentials: "include" },
-    body,
-    onSuccess,
-    onFail
-  );
-};
+} from "@/api/util";
+import { Folder, Product, User } from "@/const";
 
 export const permanentDeleteProductList = (
   productList: string[],
